requirements.txt
python-telegram-bot~=13.0
environs~=8.0.0APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.11.12
environs==8.0.0
marshmallow==3.26.1
packaging==25.0
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
tornado==6.1
tzlocal==5.3.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.11.12
environs==8.0.0
marshmallow==3.26.1
packaging==25.0
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.3.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.11.12
environs==8.0.0
marshmallow==3.26.1
packaging==25.0
psycopg2==2.9.11
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.3.1
